(@bug W-2750334@) change return for LedgerStat from int to long
BK returns long for ledgerstat, change on the wire data structure from int to long.

+++ bookkeeper-proxy/src/main/java/org/apache/bookkeeper/BKProxyWorker.java
@@ -133,12 +133,12 @@ class BKProxyWorker implements Runnable {
                     resp.put(BKPConstants.LedgerStatResp);
 
                     // Check if the extent exists
-                    if (!bksc.LedgerExists(extentIDstr)) {
+                    if (!bksc.LedgerExists(extentIDstr)) {                        
                         resp.put(BKPConstants.SF_ErrorNotFound);
                     } else {
-                        int lSize = bksc.LedgerStat(extentIDstr);
+                        long lSize = bksc.LedgerStat(extentIDstr);
                         resp.put(BKPConstants.SF_OK);
-                        resp.putInt(lSize);
+                        resp.putLong(lSize);
                     }
                     resp.flip();
                     while (resp.hasRemaining()) {
